refactor(icons): move into a dedicated folder with a barrel
Follows this repo's src/components/<Name>/<Name>.tsx + barrel
index.ts convention, which the flat icons.tsx/iconGlyphs.tsx files
didn't. index.ts's `export *` re-exports both files, restoring a
single @components/icons import path for consumers -- confirmed the
barrel doesn't retrigger react-refresh/only-export-components despite
re-exporting both a components-only and a constants-only file.
