+++ src/components/icons/index.ts
@@ -1,0 +1,2 @@
+export * from './icons'
+export * from './iconGlyphs'
