+++ src/pages/admin/RecipeEditor/RecipeEditor.tsx
@@ -11,8 +11,7 @@ import {
 import AutosaveStatus from '@components/AutosaveStatus'
 import Button from '@components/Button'
 import ConfirmDialog from '@components/ConfirmDialog'
-import { iconLock } from '@components/iconGlyphs'
-import { IconAlertCircle, IconPreview } from '@components/icons'
+import { IconAlertCircle, IconPreview, iconLock } from '@components/icons'
 import ImageUpload from '@components/ImageUpload'
 import IngredientList from '@components/IngredientList'
 import Link from '@components/Link'
