+++ src/pages/admin/UserManagement/UserManagement.tsx
@@ -2,8 +2,7 @@ import { handleSessionError } from '@api/auth'
 import { fetchUsers, inviteUser, removeUser, UserExistsError } from '@api/users'
 import Button from '@components/Button'
 import ConfirmDialog from '@components/ConfirmDialog'
-import { iconDelete, iconInvite, iconRetry, iconWarning } from '@components/iconGlyphs'
-import { IconAlertCircle } from '@components/icons'
+import { IconAlertCircle, iconDelete, iconInvite, iconRetry, iconWarning } from '@components/icons'
 import Input from '@components/Input'
 import StateBox from '@components/StateBox'
 import StatusBadge from '@components/StatusBadge'
